Made the necessary changes to the navbar
Hi, so i have made as much as possible according to the actual website. its just that the transition from left to right when you click the hamburger menu is bugging the entire code out when i'm trying so i have just done a normal dropdown menu. I can work on that in the meanwhile but for now, you can take this